Fixed compilation error when the project path included space characters
git-svn-id: http://svn.code.sf.net/p/angelscript/code/trunk@2424 404ce1b2-830e-0410-a2e2-b09542c77caf

+++ sdk/tests/test_feature/source/utils.cpp
@@ -159,7 +159,7 @@ void MyFreeWithStats(void *address)
 	numFrees++;
 
 	// Remove the memory block from the list of allocated blocks
-	int allocSize = 0;
+	size_t allocSize = 0;
 	map<void*,size_t>::iterator i = memSize.find(address);
 	if( i != memSize.end() )
 	{
